refactor: move CreeResult to search.py
